pyproject.toml
[tool.poetry]
name = "lite-spleeter"
version = "1.0.0"
description = "A very lite version of The Deezer source separation library with very low RAM usage and the same results of the original Spleeter."
authors = ["Mohammad Mansour <mohammadamansour03@gmail.com>"]
license = "MIT License"
readme = "README.md"
repository = "https://github.com/mohammadmansour200/lite-spleeter"
homepage = "https://github.com/mohammadmansour200/lite-spleeter"
classifiers = [
    "Environment :: Console",
    "Environment :: MacOS X",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Unix",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Artistic Software",
    "Topic :: Multimedia",
    "Topic :: Multimedia :: Sound/Audio",
    "Topic :: Multimedia :: Sound/Audio :: Analysis",
    "Topic :: Multimedia :: Sound/Audio :: Conversion",
    "Topic :: Multimedia :: Sound/Audio :: Sound Synthesis",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities"
]
packages = [ { include = "spleeter" } ]
include = ["LICENSE"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
ffmpeg-python = "^0.2.0"
typer = "^0.3.2"
tensorflow-cpu = "^2.5.0, <2.11.0"
pandas = "^1.3.0"
norbert = "^0.2.1"
numpy = "<2.0.0"
scipy = ">=1.8.1"

[tool.poetry.dev-dependencies]
isort = "^5.7.0"
black = "^21.7b0"
mypy = "^0.991"
flake8 = "^5.0.0"

[tool.poetry.scripts]
lite-spleeter = 'spleeter.__main__:entrypoint'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.pytest.ini_options]
addopts = "-W ignore::FutureWarning -W ignore::DeprecationWarning -vv --forked"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
